Eliminate unnecessary cast warnings; remove FileList#add(Folder) method, since it can't reliably be used to add Files that are directories (this can be done in app code).
git-svn-id: https://svn.apache.org/repos/asf/incubator/pivot/trunk@784901 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/org/apache/pivot/util/Version.java
@@ -87,10 +87,10 @@ public class Version implements Comparable<Version>, Serializable {
     }
 
     public int getNumber() {
-        int number = (((int)majorRevision) & 0xff) << (8 * 3)
-            | (((int)minorRevision) & 0xff) << (8 * 2)
-            | (((int)maintenanceRevision) & 0xff) << (8 * 1)
-            | (((int)updateRevision) & 0xff) << (8 * 0);
+        int number = ((majorRevision) & 0xff) << (8 * 3)
+            | ((minorRevision) & 0xff) << (8 * 2)
+            | ((maintenanceRevision) & 0xff) << (8 * 1)
+            | ((updateRevision) & 0xff) << (8 * 0);
 
         return number;
     }
